Trying to make signUp work
