+++ frontend/src/app/login/page.tsx
@@ -1,34 +1,10 @@
 import React, { useState } from "react";
 import { Title, Center, Group, Container, TextInput, PasswordInput, Button, Space, Stack, rem } from '@mantine/core';
 import { IconAt } from '@tabler/icons-react';
-import { signUp } from '../../../../functions/src/backend/src/endpoints/loginEndpoints.ts'
 
 export default function Login() {
-    const [loading, setLoading] = useState(false);
-    const [email, setEmail] = useState('');
-    const [displayName, setDisplayName] = useState('');
-    const [password, setPassword] = useState('');
-    const [message, setMessage] = useState('');
 
     const icon = <IconAt style={{ width: rem(16), height: rem(16) }} />;
-
-    const handleRegister = async () => {
-        setLoading(true);
-        setMessage('');
-
-        try {
-            // Firebase Authentication API call to create a new user
-            const userCredential = await signUp(email, password, displayName);
-
-            setMessage(`User registered successfully. Please check email to get verified`);
-        } catch (error: any) {
-            // Handle error and display the message
-            setMessage(`Error: ${error.message}`);
-        } finally {
-            setLoading(false);
-        }
-    };
-
 
     return (
         <Group style={{
