+++ frontend/src/app/signup/page.tsx
@@ -1,10 +1,56 @@
 'use client'
 
-import React from 'react';
+import React, { useState } from 'react';
 import { Container, Center, Title, TextInput, PasswordInput, Stack, Space, Button, rem, Group } from '@mantine/core';
 import { IconAt } from '@tabler/icons-react';
+import { getFunctions, httpsCallable } from "firebase/functions";
+
+const functions = getFunctions();
+
 
 export default function SignUp() {
+    const [loading, setLoading] = useState(false);
+    const [email, setEmail] = useState('');
+    const [displayName, setDisplayName] = useState('');
+    const [password, setPassword] = useState('');
+    const [message, setMessage] = useState('');
+
+    setEmail("random@gmail.com");
+    setDisplayName("displayname");
+    setPassword("password");
+
+    const handleRegister = async () => {
+        setLoading(true);
+        setMessage('');
+
+        try {
+            const signUp = httpsCallable(functions, 'signUp');
+            signUp({email, password, displayName})
+                .then((result) => {
+                    // Read result of the Cloud Function.
+                    /** @type {any} */
+                    const data = result.data;
+                    console.log("Data:" + data);
+                    
+                }).catch((error) => {
+                    // Getting the Error details.
+                    const code = error.code;
+                    const message = error.message;
+                    const details = error.details;
+                    setMessage(`Error: ${message}`);
+                    return;
+                    // ...
+                  });
+
+            setMessage(`User registered successfully. Please check email to get verified`);
+        } catch (error: any) {
+            // Handle any error and display the message
+            setMessage(`Error: ${error.message}`);
+        } finally {
+            setLoading(false);
+        }
+    };
+
     const icon = <IconAt style={{ width: rem(16), height: rem(16) }} />;
     const [error, setError] = React.useState('');
     const props = {
@@ -72,7 +118,7 @@ export default function SignUp() {
                         withAsterisk
                         error={error}
                     />
-                    <Button component='a' href='/edit'>Sign Up</Button>
+                    <Button component='a' href='/edit' onClick={handleRegister}>Sign Up</Button>
                 </Stack>
             </Container>
         </Group>
